% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_date_field.R
\name{add_date_field}
\alias{add_date_field}
\title{Add Date Field}
\usage{
add_date_field(df)
}
\arguments{
\item{df}{data frame; A data frame of RAS results exported from
HEC-DSSVue.}
}
\value{
The input data frame with a new POSIXct `date` field.
}
\description{
Calculates a new POSIXct `date` field built from the character
`Date` and `Time` fields exported from HEC-DSSVue.
}
