% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_files.R
\name{combine_files}
\alias{combine_files}
\title{Combine Model Results}
\usage{
combine_files(path, pattern = NULL, col_types = NULL)
}
\arguments{
\item{path}{character; Path to a folder of RAS model results.}

\item{pattern}{character; A regular expression passed to the `dir`
function to select files in the specified folder. Optional.}

\item{col_types}{readr::cols object; A column specification used to define
the columns of the input model results. Optional.}
}
\value{
A data frame containing all of the input RAS model results appended
together.
}
\description{
Combines a folder full of RAS model results in `.csv` format
into a single data frame.
}
