% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_hydrographs.R
\name{combine_hydrographs}
\alias{combine_hydrographs}
\title{Combine Hydrographs}
\usage{
combine_hydrographs(hydrograph_list)
}
\arguments{
\item{hydrograph_list}{list; A list of hydrograph event data frames.}
}
\value{
A data frame of hydrograph data with standardized column names.
}
\description{
Combines a list of hydrograph data frames and standardizes the
column names.
}
