% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gage_boxes.R
\name{gage_boxes}
\alias{gage_boxes}
\title{Create Gage Boxes}
\usage{
gage_boxes(gage_locations, stage_interval_boxes, box_width)
}
\arguments{
\item{gage_locations}{data frame; A data frame of gage locations. See
package dataset for example.}

\item{stage_interval_boxes}{numeric; The interval used for drawing tick
marks on gage boxes (units: feet).}

\item{box_width}{numeric; The width of the gage box (units:
river miles)}
}
\value{
A data frame of dimensions used by the `razviz` longitudinal profile
plot to draw gage boxes.
}
\description{
Create a data frame of gage boxes used by the longitudinal
profile plot to symbolize the location of gages and the stage elevations.
}
