% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gage_labels.R
\name{gage_labels}
\alias{gage_labels}
\title{Create Gage Labels}
\usage{
gage_labels(gage_locations, stage_interval_label)
}
\arguments{
\item{gage_locations}{data frame; A data frame of gage locations.}

\item{stage_interval_label}{numeric; The number of feet between gage stage
labels. Used to identify the location of gage
labels.}
}
\value{
A data frame of gage labels used by the `razviz` longitudinal profile
plot to label gage boxes.
}
\description{
Create a data frame of gage labels used by the longitudinal
profile plot to symbolize stage elevations.
}
