% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_stats.R
\name{gof_stats}
\alias{gof_stats}
\title{Goodness of Fit Statistics}
\usage{
gof_stats(hydrograph)
}
\arguments{
\item{hydrograph}{data frame; A hydrograph data frame in long format.}
}
\value{
A data frame of goodness of fit statistics for the between the
modeled and observed data in the input hydrograph data frame.
}
\description{
Computes a goodness of fit statistics data frame between the
modeled and observed data in the input hydrograph data frame.
}
