% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_water_events.R
\name{high_water_events}
\alias{high_water_events}
\title{High Water Events}
\usage{
high_water_events(high_water, events)
}
\arguments{
\item{high_water}{data frame; A data frame of high water marks. See the
package dataset for the required fields.}

\item{events}{character vector; A character vector of high water
event years (i.e., c("1993", "2008")).}
}
\value{
A data frame of high_water_events filtered by the `event` years.
}
\description{
Filter a data frame of high water marks for a specific list of
high water marks.
}
