% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrograph_datatables.R
\name{hydrograph_datatables}
\alias{hydrograph_datatables}
\title{Hydrograph Datatable}
\usage{
hydrograph_datatables(
  list_observed_data,
  list_1Dmodel_data,
  list_2Dmodel_data,
  locations,
  plan_events,
  run_type,
  run_number
)
}
\arguments{
\item{list_observed_data}{list; A list of data frames produced by the
`razviz::import_observed_data` function.}

\item{list_1Dmodel_data}{list; A list of data frames produced by the
`razviz::import_hecras_1Dmodel_data` function.}

\item{list_2Dmodel_data}{list; A list of data frames produced by the
`razviz::import_hecras_2Dmodel_data` function.}

\item{hg_plot_pages}{data frame; A data frame produced by the
`razviz::hydrograph_plot_pages` function.}
}
\value{
A list of dataframes.
}
\description{
Create a data table for each cross section that has observed and model data.
}
