% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrograph_plot.R
\name{hydrograph_plot}
\alias{hydrograph_plot}
\title{Hydrograph Plot}
\usage{
hydrograph_plot(plot_number, hydrograph_df, hg_plot_pages)
}
\arguments{
\item{plot_number}{numeric; The plot number to graph.}

\item{hydrograph_df}{data frame; A data frame produced by the
`razviz::lengthen_hydrographs` function.}

\item{hg_plot_pages}{data frame; A data frame produced by the
`razviz::hydrograph_plot_pages` function.}
}
\value{
A ggplot object.
}
\description{
Create a hydrograph plot.
}
