% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrograph_plot_pages.R
\name{hydrograph_plot_pages}
\alias{hydrograph_plot_pages}
\title{Hydrograph Plot Pages}
\usage{
hydrograph_plot_pages(hydrograph_df)
}
\arguments{
\item{hydrograph_df}{data frame; A data frame produced by the
`razviz::lengthen_hydrographs` function.}
}
\value{
A data frame defining the hydrograph plot pages needed for the input
hydrographs data frame.
}
\description{
Creates a data frame used to determine the number of hydrograph
plot pages needed for .
}
