% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrograph_report.R
\name{hydrograph_report}
\alias{hydrograph_report}
\title{Hydrograph Report}
\usage{
hydrograph_report(hydrograph_df, hg_plot_pages, output_dir, filename)
}
\arguments{
\item{hydrograph_df}{data frame; A data frame produced by the
`razviz::lengthen_hydrographs` function.}

\item{hg_plot_pages}{data frame; A data frame produced by the
`razviz::hydrograph_plot_pages` function.}

\item{output_dir}{character; The path to the folder where the output
report will be written.}

\item{filename}{character; The filename of the report}
}
\value{
Creates a Hydrograph Report in the output directory.
}
\description{
Create a Hydrograph Report for the input RAS model.
}
