% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_calibration.R
\name{import_calibration}
\alias{import_calibration}
\title{Import Calibration}
\usage{
import_calibration(
  folder,
  event,
  run_number,
  run_type = "Calibration",
  col_spec = NULL
)
}
\arguments{
\item{folder}{character; Path to a folder of `.csv` files of RAS model
result exports}

\item{event}{character; The name of the model event.}

\item{run_number}{numeric; The model run number.}

\item{run_type}{character; The type of model run. Label used to in plot
title.}

\item{col_spec}{readr::cols object; A column specification used to define
the columns of the input model results. Optional.}
}
\value{
A data frame of RAS model results.
}
\description{
Imports a folder of RAS model results exported using DSSVue
in `.csv` format
}
