% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv_manualrasoutput.R
\name{import_csv_manualrasoutput}
\alias{import_csv_manualrasoutput}
\title{Import csv files}
\usage{
import_csv_manualrasoutput(
  folder,
  event,
  run_number,
  run_type = "Calibration",
  col_spec = NULL
)
}
\arguments{
\item{folder}{character; Path to a folder of `.csv` files of RAS model
hydrograph `.csv` files.}

\item{event}{character; The name of the model event.}

\item{run_number}{numeric; The model run number. Used to label plot title.}

\item{run_type}{character; The type of model run. Label used to in plot
title.}

\item{col_spec}{readr::cols object; A column specification used to define
the columns of the input model results. Optional. Defaults
to the current RAS column names.}
}
\value{
A data frame of RAS model results.
}
\description{
Imports a folder of `.csv` files manually exported
from the RAS GUI. See example data for formatting.
}
