% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hecras_1Dmodel_data.R
\name{import_hecras_1Dmodel_data}
\alias{import_hecras_1Dmodel_data}
\title{Import HEC-RAS 1D Model Data}
\usage{
import_hecras_1Dmodel_data(
  model_dss_file,
  plan_names,
  plan_events,
  time_interval,
  Formatted_UnsteadyFlowFileList
)
}
\arguments{
\item{model_dss_file}{jobjRef; open dss file where observed data is stored hydrograph using dssrip.}

\item{plan_names}{list; list of the plan names as they are written in Part F of dss path.}

\item{plan_events}{list; list of years of when events used in the plans occured.}

\item{time_interval}{character; time interval chosen to run HEC-RAS model,matches Part E dss path.}

\item{Formatted_UnsteadyFlowFileList}{list; list containing the unsteady flow data for each plan;
output from import_hecras_unsteadyflowfiles function.}
}
\value{
A list of data frames with the HEC-RAS model results.
}
\description{
Imports data from dss file created from HEC-RAS model
}
