% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hecras_2Dmodel_data.R
\name{import_hecras_2Dmodel_data}
\alias{import_hecras_2Dmodel_data}
\title{Import HEC-RAS 2D Model Data}
\usage{
import_hecras_2Dmodel_data(
  model_dss_file,
  plan_names,
  plan_events,
  time_interval,
  Formatted_UnsteadyFlowFileList,
  SA2D_CONNECTIONS
)
}
\arguments{
\item{model_dss_file}{jobjRef; open dss file where observed data is stored hydrograph using dssrip.}

\item{plan_names}{list; list of the plan names as they are written in Part F of dss path.}

\item{plan_events}{list; list of years of when events used in the plans occured.}

\item{time_interval}{character; time interval chosen to run HEC-RAS model,matches Part E dss path.}

\item{Formatted_UnsteadyFlowFileList}{list; list containing the unsteady flow data for each plan;
output from import_hecras_unsteadyflowfiles function.}
}
\value{
A list of data frames with the SA Connection timeseries from HEC-RAS model results.
}
\description{
Imports 'FLOW-TOTAL'  timeseries from SA Connections in dss file created from HEC-RAS model
}
