% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hecras_unsteadyflowfiles.R
\name{import_hecras_unsteadyflowfiles}
\alias{import_hecras_unsteadyflowfiles}
\title{Import RAS Hydrographs}
\usage{
import_hecras_unsteadyflowfiles(unsteadyflowfile_list, plan_events)
}
\arguments{
\item{unsteadyflowfile_list}{list; list of dataframes which contain
unsteady flow text files}

\item{plan_events}{numeric; list of the years in which events occurred.}
}
\value{
A list data frames containing unsteady flow files.
}
\description{
Imports and formats the unsteady flow files (.u)
from the HEC-RAS GUI.
}
