% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthen_hydrographs.R
\name{lengthen_hydrographs}
\alias{lengthen_hydrographs}
\title{Lengthen Hydrographs}
\usage{
lengthen_hydrographs(hydrograph_df)
}
\arguments{
\item{hydrograph_df}{data frame; A data frame of RAS hydrograph export
data.}
}
\value{
A data frame of RAS hydrograph export data in long format.
}
\description{
Converts a data frame of RAS hydrograph export data from
wide to long format required for plotting.
}
