% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_plot_pages.R
\name{long_plot_pages}
\alias{long_plot_pages}
\title{Longitudinal Plot Pages}
\usage{
long_plot_pages(hydro_model, miles_per_plot, ..., high_water)
}
\arguments{
\item{hydro_model}{data frame; A data frame of RAS model output.}

\item{miles_per_plot}{numeric; The number of river miles per longitudinal
profile plot.}

\item{high_water}{data frame; Optional data frame of high water marks. See
package dataset `high_water_marks.csv` for format.}
}
\value{
A data frame of plot dimensions for a given set of RAS model results.
}
\description{
Creates a data frame used to define the characteristics of
each longitudinal profile plot for a given set of RAS model results. This
function determines the number of graphs needed to display a longitudinal
profile for a given RAS model. These page characteristics are used by the
`longitudinal_profile_plot` function to build each graph.
}
