% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_profile_pages.R
\name{longitudinal_profile_pages}
\alias{longitudinal_profile_pages}
\title{Longitudinal Profile Pages}
\usage{
longitudinal_profile_pages()
}
\description{
Create a data frame to represent the pages of hydrograph plots
}
