% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_event_stats.R
\name{rm_event_stats}
\alias{rm_event_stats}
\title{River Mile Event Statistics}
\usage{
rm_event_stats(hydrograph, rm_event, cal_stats)
}
\arguments{
\item{hydrograph}{data frame; A hydrograph data frame.}

\item{rm_event}{numeric; A "river mile event" created by the `gof_stats`
function.}

\item{cal_stats}{data frame; A `cal_stats` data frame produced by the
`got_stats` function. This table is used to store the
calculated goodness of fit statistics.}
}
\value{
A `cal_stats` data frame produced by the `gof_stats` function
populated with goodness of fit statistics for the specified "river mile
event".
}
\description{
Calculates goodness of fit statistics between modeled and
observed values for a single "river mile event". A "river mile event" refers
to a flood event for a specific cross section (at a specific river mile).
}
