% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_profile_plot.R
\name{longitudinal_profile_plot}
\alias{longitudinal_profile_plot}
\title{Longitudinal Profile Plot}
\usage{
longitudinal_profile_plot(
  hydro_model,
  plot_num,
  long_plot_pages,
  gages,
  gage_labels,
  gage_boxes,
  high_water,
  levees,
  features,
  bridges,
  graph_colors,
  legend_labels,
  title
)
}
\arguments{
\item{hydro_model}{data frame; A data frame of results from a RAS model. See
the package dataset for required fields.}

\item{long_plot_pages}{data frame; A data frame of longitudinal plot pages
produced by the `razviz::long_plot_pages` function.}

\item{gages}{data frame. A data frame of gages. See package dataset
for required fields.}

\item{gage_labels}{data frame; A data frame of gage labels produced by the
`razviz::gage_labels` function.}

\item{gage_boxes}{data frame; A data frame of gage boxes produced by the
`razviz::gage_boxes` function.}

\item{high_water}{data frame; A data frame of high water marks. See the
package dataset for required fields.}

\item{levees}{data frame; A data frame of levee elevations. See the
package dataset for required fields.}

\item{features}{data frame; A data frame of salient river features (e.g.,
tributary confluences, cities, bridges, dams, etc.). See
package dataset for required fields.}

\item{bridges}{data frame; A data frame of bridge opening elevations.
See package dataset for required fields.}

\item{graph_colors}{character vector; A named character vector of colors to
use for each series.}

\item{title}{character; A title for the plot.}

\item{plot_number}{numeric; The plot number to graph.}

\item{legend_label}{character vector; A character vector of labels to be
used in the legend.}
}
\value{
A `ggplot2` object depicting the river longitudinal profile graph.
}
\description{
Creates a longitudinal profile graph for the input RAS
hydro model.
}
